Update endpoint-antimalware-policies.md
The **User control of scheduled scans** seems to work only for SCEP and not Windows Defender. I would like to see this reflected in our documentation so our customers know this. Also, if we can add more details of what No control, Scan only and full control means and what capabilities we have with each option, it will be nice.